+++ src/mantle/framework/console/class-kernel.php
@@ -18,11 +18,14 @@ use Mantle\Contracts\Exceptions\Handler as Exception_Handler;
 use Mantle\Support\Traits\Loads_Classes;
 use ReflectionClass;
 use Symfony\Component\Console\Exception\CommandNotFoundException;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\NullOutput;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Tester\CommandTester;
 use Throwable;
 
 use function Mantle\Support\Helpers\collect;
+use function Mantle\Support\Helpers\stringable;
 
 /**
  * Console Kernel
@@ -77,14 +80,14 @@ class Kernel implements \Mantle\Contracts\Console\Kernel {
 	/**
 	 * Run the console application
 	 *
-	 * @param \Symfony\Component\Console\Input\InputInterface   $input
-	 * @param \Symfony\Component\Console\Output\OutputInterface $output
+	 * @param InputInterface|null  $input Console input.
+	 * @param OutputInterface|null $output Console output.
 	 */
-	public function handle( $input = null, $output = null ): int {
+	public function handle( ?InputInterface $input = null, ?OutputInterface $output = null ): int {
 		$this->output = $output;
 
 		try {
-			return $this->get_console_application()->run( $input, $output );
+			return $this->get_console_application()->run( $input, $output ?? $this->output );
 		} catch ( Throwable $e ) {
 			$this->report_exception( $e );
 			$this->render_exception( $output, $e );
@@ -96,14 +99,24 @@ class Kernel implements \Mantle\Contracts\Console\Kernel {
 	/**
 	 * Run the console application by command name.
 	 *
-	 * @param string $command Command name.
-	 * @param array  $parameters Command parameters.
-	 * @param mixed  $output_buffer Output buffer.
+	 * @param string               $command Command name.
+	 * @param array                $parameters Command parameters.
+	 * @param OutputInterface|null $output_buffer Output buffer.
 	 */
-	public function call( string $command, array $parameters = [], $output_buffer = null ): int {
+	public function call( string $command, array $parameters = [], ?OutputInterface $output_buffer = null ): int {
 		$this->bootstrap();
 
-		return $this->get_console_application()->call( $command, $parameters, $output_buffer );
+		return $this->get_console_application()->call( $command, $parameters, $output_buffer ?? $this->output );
+	}
+
+	/**
+	 * Run the console application by command name without output.
+	 *
+	 * @param string $command Command name.
+	 * @param array  $parameters Command parameters.
+	 */
+	public function call_silently( string $command, array $parameters = [] ): int {
+		return $this->call( $command, $parameters, new NullOutput() );
 	}
 
 	/**
@@ -276,7 +289,7 @@ class Kernel implements \Mantle\Contracts\Console\Kernel {
 	 */
 	protected function render_exception( OutputInterface $output, Throwable $e ) {
 		if ( $e instanceof CommandNotFoundException ) {
-			$this->output->writeln( '<error>' . str( $e->getMessage() )->explode( '.' )->first() . '</error>' );
+			$this->output->writeln( '<error>' . stringable( $e->getMessage() )->explode( '.' )->first() . '</error>' );
 			$this->output->writeln( '' );
 
 			if ( ! empty( $alternatives = $e->getAlternatives() ) ) {