+++ src/mantle/framework/exceptions/class-handler.php
@@ -9,6 +9,7 @@ namespace Mantle\Framework\Exceptions;
 
 use Exception;
 use Mantle\Auth\Authentication_Error;
+use Mantle\Console\Manually_Failed_Exception;
 use Mantle\Contracts\Application;
 use Mantle\Contracts\Exceptions\Handler as Contract;
 use Mantle\Database\Model\Model_Not_Found_Exception;