+++ CHANGELOG.md
@@ -5,17 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## v1.5.6
 
 ### Added
 
 - Added a `mixed()` helper function to return a `Interacts_With_Data` instance
   for a mixed value that isn't associated with meta or an option.
 
+### Changed
+
+- Updated `block_factory()->block()` to require `null` to be passed as `$content`
+  if no content is desired. This prevents the block factory from generating a
+  an empty block when `$content` is just an empty string.
+
 ### Fixed
 
 - Prevent the site's default category from being added to a post when using the
   `with_terms()` method on a post factory if a category is being set.
+- Fixed an issue where the output would not pass back to the console when
+  calling a command with `Console::call()`.
 
 ## v1.5.5 - 2025-03-12